+++ openlane/__main__.py
@@ -30,9 +30,6 @@ from cloup import (
     Context,
 )
 from cloup.constraints import (
-    If,
-    IsSet,
-    accept_none,
     mutually_exclusive,
 )
 
@@ -256,6 +253,7 @@ def cli_in_container(
 
     status = 0
     docker_mounts = list(ctx.params.get("docker_mounts") or ())
+    docker_tty: bool = ctx.params.get("docker_tty", True)
     pdk_root = ctx.params.get("pdk_root")
     argv = sys.argv[sys.argv.index("--dockerized") + 1 :]
 
@@ -265,13 +263,18 @@ def cli_in_container(
         final_argv = ["openlane"] + argv
         interactive = False
 
+    docker_image = os.getenv(
+        "OPENLANE_IMAGE_OVERRIDE", f"ghcr.io/efabless/openlane2:{__version__}"
+    )
+
     try:
         status = run_in_container(
-            f"ghcr.io/efabless/openlane2:{__version__}",
+            docker_image,
             final_argv,
             pdk_root=pdk_root,
             other_mounts=docker_mounts,
             interactive=interactive,
+            tty=docker_tty,
         )
     except ValueError as e:
         print(e)
@@ -297,19 +300,24 @@ o = partial(option, show_default=True)
         "-m",
         "docker_mounts",
         multiple=True,
-        is_eager=True,  # docker mount, dockerized should be processed before anything else
-        default=[],
-        help="Additionally mount this directory in dockerized mode. Can be supplied multiple times to mount multiple directories. Must be passed before --dockerized.",
+        is_eager=True,  # docker options should be processed before anything else
+        default=(),
+        help="Used to mount more directories in dockerized mode. If a valid directory is specified, it will be mounted in the same path in the container. Otherwise, the value of the option will be passed to the Docker-compatible container engine verbatim. Must be passed before --dockerized, has no effect if --dockerized is not set.",
+    ),
+    o(
+        "--docker-tty/--docker-no-tty",
+        is_eager=True,  # docker options should be processed before anything else
+        default=True,
+        help="Controls the allocation of a virtual terminal by passing -t to the Docker-compatible container engine invocation. Must be passed before --dockerized, has no effect if --dockerized is not set.",
     ),
     o(
         "--dockerized",
         default=False,
         is_flag=True,
-        is_eager=True,  # docker mount, dockerized should be processed before anything else
-        help="Re-invoke using a Docker container. Some caveats apply. Must precede all options except --docker-mount.",
+        is_eager=True,  # ddocker options should be processed before anything else
+        help="Run the remaining flags using a Docker container. Some caveats apply. Must precede all options except --docker-mount, --docker-tty/--docker-no-tty.",
         callback=cli_in_container,
     ),
-    constraint=If(~IsSet("dockerized"), accept_none),
 )
 @option_group(
     "Subcommands",
@@ -353,6 +361,7 @@ def cli(ctx, /, **kwargs):
         run_kwargs.update(**{k: kwargs[k] for k in ["pdk_root", "pdk", "scl"]})
 
     for subcommand_flag in [
+        "docker_tty",
         "docker_mounts",
         "dockerized",
         "version",